# App-Function/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GlobalMarket API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: region
          in: query
          schema:
            type: string
        - name: sellerId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
...
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List my orders (as buyer or seller)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [buyer, seller]
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order (buyer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created (in escrow)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
...
components:
...
  schemas:
...
    Product:
      type: object
      required: [id, name, price, currency, country, category, featured, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        currency:
          type: string
        country:
          type: string
        category:
          type: string
        featured:
          type: boolean
        imageUrl:
          type: ["string", "null"]
        sellerId:
          type: ["string", "null"]
        sellerName:
          type: ["string", "null"]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, price, currency, country, category]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
        country:
          type: string
        category:
          type: string
        featured:
          type: boolean
        imageUrl:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        currency:
          type: string
        country:
          type: string
        category:
          type: string
        featured:
          type: boolean
        imageUrl:
          type: string
...
    Order:
      type: object
      required: [id, productId, buyerId, sellerId, price, currency, status, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        buyerId:
          type: string
        sellerId:
          type: string
        price:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, paid, shipped, delivered, disputed, cancelled]
        productName:
          type: ["string", "null"]
        productImageUrl:
          type: ["string", "null"]
        buyerName:
          type: ["string", "null"]
        sellerName:
          type: ["string", "null"]
        createdAt:
          type: string
        shippedAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
    OrderInput:
      type: object
      required: [productId]
      properties:
        productId:
          type: integer